requirements/base.txt
Django==1.7.1
djangorestframework==3.0.3
